Fixup: unify create param models, change location of should calibrate tip length
