+++ robot-server/robot_server/service/session/models/session.py
@@ -9,12 +9,12 @@ from robot_server.robot.calibration.check.models import \
     CalibrationSessionStatus
 from robot_server.robot.calibration.deck.models import \
     DeckCalibrationSessionStatus
-from robot_server.robot.calibration.pipette_offset.models import \
-    SessionCreateParams as PipetteOffsetSessionCreateParams, \
+from robot_server.robot.calibration.models import \
+    SessionCreateParams as TipLengthPipetteOffsetSessionCreateParams
+from robot_server.robot.calibration.pipette_offset.models import\
     PipetteOffsetCalibrationSessionStatus
-from robot_server.robot.calibration.tip_length.models import (
-    SessionCreateParams as TipLengthSessionCreateParams,
-    TipCalibrationSessionStatus)
+from robot_server.robot.calibration.tip_length.models import\
+    TipCalibrationSessionStatus
 from robot_server.service.json_api import RequestModel, ResponseModel
 from robot_server.service.json_api.response import MultiResponseModel
 from robot_server.service.session.models.common import EmptyModel
@@ -42,12 +42,12 @@ class SessionType(str, Enum):
     calibration_check = 'calibrationCheck'
     tip_length_calibration = (
         'tipLengthCalibration',
-        TipLengthSessionCreateParams
+        TipLengthPipetteOffsetSessionCreateParams
     )
     deck_calibration = 'deckCalibration'
     pipette_offset_calibration = (
         'pipetteOffsetCalibration',
-        PipetteOffsetSessionCreateParams
+        TipLengthPipetteOffsetSessionCreateParams
     )
     protocol = ('protocol', ProtocolCreateParams)
     live_protocol = 'liveProtocol'
@@ -69,8 +69,7 @@ When we move to Python 3.8 we can use Literal type as described here
 https://pydantic-docs.helpmanual.io/usage/types/#literal-type
 """
 SessionCreateParamType = typing.Union[
-    PipetteOffsetSessionCreateParams,
-    TipLengthSessionCreateParams,
+    TipLengthPipetteOffsetSessionCreateParams,
     ProtocolCreateParams,
     None,
     EmptyModel
